+++ src/main/java/com/vandendaelen/k9/gui/K9ContainerGui.java
@@ -1,8 +1,8 @@
 package com.vandendaelen.k9.gui;
 
 import com.vandendaelen.k9.objects.containers.K9Container;
-import com.vandendaelen.k9.objects.tilesentities.K9ContainerTileEntity;
 import com.vandendaelen.k9.utils.Reference;
+
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.util.ResourceLocation;
 
@@ -12,7 +12,7 @@ public class K9ContainerGui extends GuiContainer {
 
     private static final ResourceLocation background = new ResourceLocation(Reference.MODID, "textures/gui/k9container.png");
 
-    public K9ContainerGui(K9ContainerTileEntity tileEntity, K9Container container) {
+    public K9ContainerGui(K9Container container) {
         super(container);
 
         xSize = WIDTH;
